Remove support for global dtype
