--- geomstats/_backend/_backend_config.py
@@ -1,9 +1,0 @@
-pytorch_atol = 1e-6
-pytorch_rtol = 1e-5
-
-np_atol = 1e-8
-np_rtol = 1e-5
-
-
-DEFAULT_DTYPE = None
-DEFAULT_COMPLEX_DTYPE = None
