+++ geomstats/_backend/numpy/random.py
@@ -4,4 +4,13 @@ import numpy as _np
 from numpy.random import default_rng as _default_rng
 from numpy.random import randint, seed
 
-from .._shared_numpy.random import choice, multivariate_normal, normal, rand, uniform
+from ._dtype import _allow_complex_dtype
+
+rand = _allow_complex_dtype(target=_np.random.rand)
+uniform = _allow_complex_dtype(target=_np.random.uniform)
+normal = _allow_complex_dtype(target=_np.random.normal)
+multivariate_normal = _allow_complex_dtype(target=_np.random.multivariate_normal)
+
+
+def choice(*args, **kwargs):
+    return _np.random.default_rng().choice(*args, **kwargs)
